Adding missing code from main
