+++ src/test/java/org/opensearch/searchrelevance/action/judgment/JudgmentsIT.java
@@ -49,7 +49,7 @@ public class JudgmentsIT extends BaseSearchRelevanceIT {
         assertNotNull(judgmentsId);
 
         // wait for completion of import action
-        Thread.sleep(1000);
+        Thread.sleep(DEFAULT_INTERVAL_MS);
 
         String getJudgmentsByIdUrl = String.join("/", JUDGMENT_INDEX, "_doc", judgmentsId);
         Response getJudgmentsResponse = makeRequest(
