+++ src/test/resources/experiment/CreateExperimentHybridOptimizer.json
@@ -1,7 +1,7 @@
 {
-  "querySetId": "{{query_set_id}}",
-  "searchConfigurationList": ["{{search_config_id}}"],
-  "size": 2,
-  "judgmentList": ["{{judgment_id}}"],
-  "type": "HYBRID_OPTIMIZER"
+    "querySetId": "{{query_set_id}}",
+    "searchConfigurationList": ["{{search_config_id}}"],
+    "size": 2,
+    "judgmentList": ["{{judgment_id}}"],
+    "type": "HYBRID_OPTIMIZER"
 }
