+++ src/test/resources/searchconfig/CreateSearchConfigurationHybridQuery.json
@@ -1,6 +1,5 @@
 {
     "name": "hybrid_query_1",
-    "index": "ecommerce",
-    "query": "{\"query\":{\"hybrid\":{\"queries\":[{\"match\":{\"title\":\"%SearchText%\"}},{\"match\":{\"category\":\"%SearchText%\"}}]}}}",
-    "searchPipeline": ""
+    "index": "{{index_name}}",
+    "query": "{\"query\":{\"hybrid\":{\"queries\":[{\"match\":{\"title\":\"%SearchText%\"}},{\"match\":{\"category\":\"%SearchText%\"}}]}}}"
 }
